\name{spatcounts-package}
\alias{spatcounts-package}
\alias{spatcounts}
\docType{package}
\title{
Spatial count regression
}
\description{
Fit spatial CAR count regression models using MCMC.
}
\details{
\tabular{ll}{
Package: \tab spatcounts\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-06-08\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
}
Spatial data can be fit by MCMC using function 'est.sc'. Functions 
'Vuongtest' and 'Clarketest' allow for nonnested model comparison, 
'DIC' calculates the Deviance Information Criterion, 'LogLike' 
calculates the loglikelihoods.
}
\author{
Holger Schabenberger

Maintainer: Vinzenz Erhardt <erhardt@ma.tum.de>
}
\references{
Masterthesis: 
Schabenberger, Holger (2009). Spatial count regression models with applications 
to health insurance data. ("http://www-m4.ma.tum.de/Diplarb/").
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
\keyword{ package }
\keyword{ spatial }
\keyword{ regression }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
